fix : Create component for home charts
