+++ src/components/home/HostMostVulnerabilitiesList.vue
@@ -1,0 +1,39 @@
+<template>
+  <div class="title">Hosts with the most vulnerabilities</div>
+
+  <q-list>
+    <q-item>
+      <q-item-section avatar></q-item-section>
+      <q-item-section class="text-weight-bold label-table">TOP 5</q-item-section>
+      <q-item-section class="text-weight-bold label-table"> # OF VULNERABILITIES</q-item-section>
+    </q-item>
+    <template v-for="i in hostVulnerabilities" :key="i">
+      <q-item v-ripple clickable>
+        <q-item-section avatar>
+          <i class="fa-solid fa-square q-mx-sm" :style="`color:${i.color};font-size:2em`"></i>
+        </q-item-section>
+
+        <q-item-section>{{ i.alias }}</q-item-section>
+        <q-item-section>{{ i.vulnerability_count }}</q-item-section>
+      </q-item>
+    </template>
+  </q-list>
+</template>
+
+<script setup lang="ts">
+  import type { iMainDashboard } from 'src/models/dashboard.models';
+  import { MainDashboard } from 'src/services/dashboard';
+  import { computed } from 'vue';
+  import type { PropType } from 'vue';
+
+  const props = defineProps({
+    dashboard: {
+      type: Object as PropType<iMainDashboard>,
+      required: true
+    }
+  });
+
+  const hostVulnerabilities = computed(
+    () => new MainDashboard(props.dashboard).listHostsWithGreatestVulnerabilities
+  );
+</script>
