+++ src/components/home/LatestScanChart.vue
@@ -1,0 +1,67 @@
+<template>
+  <div class="row q-mb-md">
+    <div class="col-12">
+      <div class="title">Latest Scan</div>
+      <div class="subtitle" style="color: #5c7288">Vulnerability Count</div>
+      <div class="subtitle" style="color: #5c7288">
+        <span style="font-weight: 500"> Date :</span>
+        {{ dashboard.last_scan?.scan_date.slice(0, 10) || '' }}
+      </div>
+    </div>
+
+    <div class="full-width">
+      <template v-if="lastScanEmpty">
+        <div class="subtitle" style="color: #5c7288">
+          No scan data to display. Please run a scan to see details.
+        </div>
+      </template>
+      <template v-else>
+        <apexchart
+          :options="SCAN_INSIGHT_VULNERABILITY_OPTIONS"
+          :series="donutSeverityCountsSeries"
+        ></apexchart>
+      </template>
+    </div>
+  </div>
+  <div class="row flex items-center q-mb-md protection">
+    <div class="col-6 flex items-center justify-center text-weight-bold protection-score">
+      {{ dashboard?.last_scan?.total_vulnerabilities }}
+    </div>
+    <div class="col-6">
+      <p class="q-ma-none protection-host">
+        {{ dashboard?.last_scan?.host_alias }}
+      </p>
+      <p class="q-ma-none protection-variation">
+        {{ dashboard?.last_scan?.total_vulnerabilities_variation }}
+        <i :class="variationIcon"></i>
+        Variation
+      </p>
+    </div>
+  </div>
+</template>
+
+<script lang="ts" setup>
+  import type { iMainDashboard } from 'src/models/dashboard.models';
+  import { MainDashboard } from 'src/services/dashboard';
+  import type { ComputedRef, PropType } from 'vue';
+  import { getVariationIcon } from 'src/utils';
+  import { computed } from 'vue';
+  import { SCAN_INSIGHT_VULNERABILITY_OPTIONS } from 'src/constants/apexcharts.constants';
+
+  const props = defineProps({
+    dashboard: {
+      type: Object as PropType<iMainDashboard>,
+      required: true
+    }
+  });
+
+  const donutSeverityCountsSeries = computed(
+    () => new MainDashboard(props.dashboard).donutSeverityCountsSeries
+  );
+
+  const variationIcon: ComputedRef<string> = computed(() =>
+    getVariationIcon(props.dashboard.overall_security_posture?.variation || 0)
+  );
+
+  const lastScanEmpty = computed(() => donutSeverityCountsSeries.value.every(val => val === 0));
+</script>
