+++ src/components/home/VulnerabilityHeatMap.vue
@@ -1,0 +1,24 @@
+<template>
+  <apexchart type="heatmap" :options="heatmapOptions" :series="heatmapSeries"></apexchart>
+</template>
+
+<script lang="ts" setup>
+  import { computed, type PropType } from 'vue';
+  import { HEATMAP_CHART_OPTIONS } from 'src/constants/apexcharts.constants';
+  import type { iMainDashboard } from 'src/models/dashboard.models';
+  import { MainDashboard } from 'src/services/dashboard';
+
+  const props = defineProps({
+    dashboard: {
+      type: Object as PropType<iMainDashboard>,
+      required: true
+    }
+  });
+
+  const heatmapCategories = computed(() => new MainDashboard(props.dashboard).heatmapCategories);
+  const heatmapSeries = computed(() => new MainDashboard(props.dashboard).heatmapSeries);
+  const heatmapOptions = computed(() => ({
+    ...HEATMAP_CHART_OPTIONS,
+    xaxis: { categories: heatmapCategories.value, type: 'category' }
+  }));
+</script>
